requirements.txt
argon2-cffi==20.1.0
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-resized==0.3.11
django-simple-history==2.12.0
gunicorn==20.1.0
Pillow==8.2.0
# psycopg2==2.8.6
psycopg2-binary==2.9.1
pytz==2020.4
sqlparse==0.4.1
twilio==6.52.0
boto3==1.17.69
django-storages==1.11.1
xlrd==1.2.0
xlwt==1.3.0
openpyxl